requirements.txt
asgiref==3.11.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-extensions==4.1
django-jazzmin==3.0.1
django-nested-admin==4.1.6
django-redis==6.0.0
fontawesomefree==6.6.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gunicorn==23.0.0
httplib2==0.31.0
idna==3.11
oauthlib==3.3.1
packaging==25.0
pillow==12.1.0
proto-plus==1.27.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.3.1
pypdf==6.5.0
python-decouple==3.8
python-dotenv==1.2.1
python-magic==0.4.27
python-magic-bin==0.4.14
python-monkey-business==1.1.0
redis==7.1.0
reportlab==4.4.7
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.2
whitenoise==6.11.0
